% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{addMajorMinorAllele}
\alias{addMajorMinorAllele}
\title{Add Minor and Major allele ploidy levels}
\usage{
addMajorMinorAllele(sample_data)
}
\arguments{
\item{sample_data}{Data frame with all information from the .cnv file}
}
\value{
sample_data
}
\description{
Adds major and minor allele ploidy data to the data frame
}
\examples{
addMajorMinorAllele(sample_data)
}
