% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{aneuPloidy}
\alias{aneuPloidy}
\title{Determine ploidy level of chromosome arm}
\usage{
aneuPloidy(ploidyFrame, chromosome, arm, filename)
}
\arguments{
\item{ploidyFrame}{Data frame with segment information of one chromosome arm}

\item{chromosome}{Integer that indicates which chromosome is represented in arm_subset}

\item{arm}{Character that indicates which chromosome arm is represented in arm_subset}

\item{filename}{Character string of the file that need to be read.}
}
\value{
list with major and minor allele ploidy levels
}
\description{
Determines ploidy level of the autosome arms
}
\examples{
aneuPloidy(ploidyFrame, 12 , "p", "example.cnv")
}
