% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{chromArmPloidy}
\alias{chromArmPloidy}
\title{Determine ploidy levels}
\usage{
chromArmPloidy(sample_data, filename)
}
\arguments{
\item{sample_data}{Data frame with all information from the .cnv file}

\item{filename}{Character string of the file that need to be read.}
}
\value{
list with major and minor ploidy levels per autosome arm
}
\description{
Calculates per chromosome arm the average major and minor ploidy (weighted)
}
\examples{
chromArmPloidy(sample_data, "example.cnv")
}
