% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainPloidyEstimator.R
\name{main}
\alias{main}
\title{Main}
\usage{
main()
}
\value{
nothing but does write PDF files with all plots of a file and a summary pdf of all files.
}
\description{
Main function which calls all necessary functions for determination of ploidy levels per autosome
}
