% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{optimalization}
\alias{optimalization}
\title{Optimalize the major and minor ploidy levels}
\usage{
optimalization(ploidy_list, filename, tp)
}
\arguments{
\item{ploidy_list}{list of ploidy levels}

\item{filename}{Character string of the file that need to be read.}

\item{tp}{Character string that defines if major or minor doubt cases get normalized}
}
\value{
ploidy_list
}
\description{
Determines overall MajorAP level of the genome and changes the NA values accordingly with previous stores doubt_df data.
}
\examples{
optimalization(ploidy_list, "example.cnv", "major")
}
