% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{ploidyWeight}
\alias{ploidyWeight}
\title{Add ploidy weight}
\usage{
ploidyWeight(data_frame)
}
\arguments{
\item{data_frame}{Data frame with all information from the .cnv file}
}
\value{
data_frame
}
\description{
Adds the right weight score to the data frame
}
\examples{
ploidyWeight(data_frame)
}
