% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{readFile}
\alias{readFile}
\title{Reading .cnv file}
\usage{
readFile(filename)
}
\arguments{
\item{filename}{Character string of the file that need to be read.}
}
\value{
sample_data
}
\description{
Reads file and makes data frame of data (sample_data)
}
\examples{
readFile("example.cnv")
}
