% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{smoothing}
\alias{smoothing}
\title{Combine segments based on ploidy level}
\usage{
smoothing(arm_subset, chromosome, arm)
}
\arguments{
\item{arm_subset}{Data frame with segment information of one chromosome arm}

\item{chromosome}{Integer that indicates which chromosome is represented in arm_subset}

\item{arm}{Character that indicates which chromosome arm is represented in arm_subset}
}
\value{
ploidyFrame
}
\description{
Creates new table with combined segments if it's possible
}
\examples{
smoothing(arm_subset, 13 , "q")
}
